Add fissionable check to pointwise XS lookups
In the main branch of OpenMC, we have a check when performing pointwise XS lookups to just set `fission` and `nu_fission` to zero if `fissionable_` is false. In our offloading branch, this test isn't present. I'm not sure if this can cause issues, or if the underlying nuclide fission/nu_fission data is zero already, but worth adding this in to maintain the same logic.